+++ modules/@apostrophecms/modal/ui/apos/components/AposModalShareDraft.vue
@@ -4,7 +4,6 @@
     class="apos-share-draft"
     data-apos-test="share-draft-modal"
     v-on="{ esc: close }"
-    @no-modal="$emit('safe-close')"
     @inactive="modal.active = false"
     @show-modal="modal.showModal = true"
   >
@@ -88,7 +87,6 @@ export default {
       required: true
     }
   },
-  emits: [ 'safe-close' ],
   data() {
     return {
       modal: {